+++ implementation/js/src/infrastructure/lib+-/underscore/custom-apis.js
@@ -5,11 +5,16 @@
  * @created 2017.02.04
  */
 
-(function(_, underscoreString){
+(function(_, underscoreString, $){
 
 	_.isPlainObject = function(o){
 		return _.isObject(o) && !_.isFunction(o) && !_.isArray(o) && !_.isElement(o);
 	};
+
+	_.isjQueryObject = function(o){
+		return o instanceof $;
+	};
+
 	_.string = underscoreString;
 
-})(_, s);
+})(_, s, jQuery);